Refactored code for even handling in adapters
